pyproject.toml
[project]
name = "classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mdop297", email = "lenhatminh297@gmail.com" }
]

requires-python = ">=3.10, <3.11"
dependencies = [
    "facenet-pytorch>=2.5.3",
    "matplotlib>=3.9.0",
    "mediapipe==0.10.9",
    "numpy>=1.26.0,<2.0.0",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "seaborn>=0.13.2",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm>=4.66.0",
    "pillow>=10.0.0",
    "mlflow>=3.8.0",
    "protobuf==3.20.3",
]

[project.scripts]
classifier = "classifier:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
